+++ airdnb/src/main/java/team10/airdnb/oauth/kakao/client/KakaoTokenClient.java
@@ -1,9 +1,10 @@
-package team10.airdnb.oauth.kakao;
+package team10.airdnb.oauth.kakao.client;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.cloud.openfeign.SpringQueryMap;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
+import team10.airdnb.oauth.kakao.dto.KakaoTokenDto;
 
 @FeignClient(url = "https://kauth.kakao.com", name = "kakaoTokenClient")
 public interface KakaoTokenClient {
