+++ airdnb/src/main/java/team10/airdnb/oauth/kakao/controller/KakaoTokenController.java
@@ -1,10 +1,12 @@
-package team10.airdnb.oauth.kakao;
+package team10.airdnb.oauth.kakao.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import team10.airdnb.oauth.kakao.dto.KakaoTokenDto;
+import team10.airdnb.oauth.kakao.client.KakaoTokenClient;
 
 @Controller
 @RequiredArgsConstructor
